Move gRPC helper functions and types to `rpc` mod.
Make them more accessible so that they can be used by other tests.
